candidates/2026-03-17/requirements.txt
google-genai>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
click>=8.1.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
